Replay logblock allocation log
We have to reserve logblock as allocated. sb->logchain's block address
has address to do it. We use it like balloc log and update bitmap in
replay stage2.

FIXME: Now, this is using address array to know logblock address,
because logchain is reverse order. We would want to rewrite it with
more efficient way, instead reparse logchain for each stages.

We will revisit later if atomic-commit is more completed.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>