Add LOG_ROLLUP and LOG_DELTA log
For now, We have to know where is the log that rollup was already done
or not. To know it, this adds LOG_ROLLUP. And just for debugging,
LOG_DELTA is added (even if delta was empty, we will write
LOG_DELTA.).

LOG_ROLLUP is an initial log on rollup.
LOG_DELTA is an initial log on delta.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>