Cursor series, make cursor->len to cursor->level as level of cursor position
cursor->len was born as used counts of path_level array of
cursor->path, then become to be used as cursor position. And it was
useful.

To use it as cursor position more, this rename ->len to ->level. And
->len has different base with btree->depth, it was confusable to
use. So this adds "->level" with "->len - 1". With it, we can compare
cursor->level easily with btree->depth.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>