Rewrite bufvec API based on the page
In the kernel, we have to handle the flags on page. To do it, we need
to control BIOs based on the page.

If the page across on multiple logical ranges (non-contiguous), we
will submit multiple BIOs for the page. In this case, to change page
state (e.g. writeback flag), we have to make sure all BIOs was done
for the page.

So, this teaches about the page (page boundary) to bufvec. As result,
this rewrites bufvec stuff.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>