Rewrite bufvec API based on kernel bufvec
Current userland bufvec API was having unnecessary limitation of
maximum contiguous range. We can avoid it by using list, instead of
preallocated array.

New API uses buffer->link for candidate buffers on contiguous range.
NOTE, this means we have to re-add to proper buffer state list from
bufvec, this would be main disadvantage.

With this change, bufvec has all candidate buffers on the contiguous
range. It guarantees next candidate buffer is not contiguous, so the
block allocator doesn't have to care about next buffer which is yet
not in bufvec.

And for future, this introduces I/O completion interface, maybe for
asynchronous.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>